Separating sorting functions from multi_column_sort file. Signed-Off-By: Piyush Vijay <piyushvijay.1997@gmail.com>
